removed macro argument name collision in MAKEDATA in different place, 105 total

+++ XmlScannerPseudoMacro.h
@@ -162,7 +162,7 @@ static int scanXml(
 //	NSLog(@"start scan with client data %x",clientData);
     while ( currentPtr < endPtr ) {
         const xmlchar *currentString = currentPtr;
-        int spaceOffset=0;
+        long spaceOffset=0;
         ProcessFunc currentCallback;
 
 		//--- scan up to the beginning of a tag (the initial '<' )
